Prefetch layers into vram
In principle I would expect this to be faster, but I'm not actually
seeing any difference in performance from the previous version.

Potentially PCI bandwidth is just not a problem at all? Even a few
percent satuartion should show up as the previous version was
effectively blocking on (nearly) all the memory movement.